Filter out trivial `let_` expressions.
I think `let_` is probably useless. The only situation I can think of
where `let_` won't be filtered out is if your base terms include
something like a function applied to a variable, as variable
assignment is just not done.

The problem with just removing `let_` and using assignment is that
it's less easy to predict how the size of the expression will change.