Typey equivalence classes.
Consider this:

```
takeMVar :state:  is equivalent to  void (takeMVar :state:)
````

It's true, as there is exactly the same effect on the state. However,
this does not mean (unlike in QuickSpec or Speculate) that we can
throw away the larger term and only use the smaller in generating new
terms, as they have different types!

What we actually need to do is keep both around, if the types differ,
but then do some pruning to avoid testing cases where it doesn't
matter. In our use-case, it only matters in a polymorphic bind
(`>>= \_ -> ...`). So we can do so:

- If two terms have the same type, throw away the larger.
- Otherwise, annotate the larger, saying that it is not the smallest
  equivalent expression.
- When producing new expressions:
    - If it's a bind: the right-hand side must be the smallest, and
      the left-hand side must be the smallest if it's
      polymorphic. Otherwise, throw it away.
    - Optimistically assume that all new expressions are the smallest
      in their equivalence class.
- When testing expressions, only compare pairs of smallests.

This means that the `Ann` type is no longer a monoid, as if combining
is optimistic about size, there is no longer a unit value.