Cache all concurrent execution results.
This mostly eliminates duplicate runs of terms. When discovering
properties on tier N, all the terms on tiers <N will not be run again,
their results are already known.

BUT: the terms on tier N will still be run many times, sadly. Fixing
that will come in a separate commit.