Strictify data for performance.
This makes things slightly faster, and it makes sense: a load of data
is constructed, inspected, and then printed (in the executables). It
all gets forced anyway!

Note: Test.Spec.Type is *not* using StrictData, because it actually
slows things down a little! I find this pretty surprising. How odd.