Introduce a notion of "boring" terms.
Boring terms are nonfailing, atomic, and have no effect on the state:
this means that they are all equivalent, and when composed with
another term will not affect its result.

Not yet implemented: marking terms as boring after evaluating them.

In the future, a notion of boring failing terms may be interesting as
well.