Constrain the seed and observation types to be the same.
This is needed to check if the state has been changed at all. Another
option would be to "test" a constructed no-op term for each seed
value, but I realised I wasn't making use of the different types in
either of the examples, and I'm not sure where it would be
inconvenient/difficult anyway.