Fix reporting of false refinements.
The results-with-interference are absolutely necessary to distinguish
equivalences from refinements in the lack of additional
concurrency. However, they are actively misleading when just finding
refinements! It is true but totally uninteresting that
with *identical* interference, many things are refinements of each
other.

To resolve this:

1. Produce refinement/equivalence candidates by looking at the
   results-without-interference
2. Potentially turn equivalences into refinements by looking at the
   results-with-interference.

This has the downside of increasing the amount of computation done on
each term. It's not a doubling, because adding one thread does more
than double the number of potential executions, but it's still a lot.