Introduce a function to avoid boring commutativity properties
In many cases, we know that a function is commutative. For example,
the two parallel composition operators provided:

   (|||) :: m a -> m b -> m ()
   (|+|) :: m a -> m b -> m ()

A lot of spurious properties get discovered, which only differ from
each other by argument order to these commutative functions.

By introducing a `commutativeConstant` function, such
known-commutative functions can be handled specially, by making
function application fail (as if the application were
improperly-typed) if the arguments are not applied in a canonical
ordering, which is currently just `<` for `Expr`.

This is probably only useful for background functions, where
commutativity is a boring property. For non-background functions,
commutativity may be unexpected and interesting!

This change has a ripple effect. By forbidding some terms, many larger
ones are not generated. This reduces execution time of the examples
significantly:

        MVar   Stack   ComplexStack
Before:  59s    126s           749s
After:   34s     68s           401s

(times measures on azathoth)