Project term variables into a consistent namespace.
See this memo for a fuller discussion:

- https://memo.barrucadu.co.uk/mixing-and-matching-variables.html

The intent is to solve the problem of variable names only making sense
within a single term, effectively being namespaced to that scope.

Say we have two terms

```
term1 = f (x :: Int) (y :: Int)  (z :: Bool)
term2 = g (x :: Int) (y :: Bool) (z :: Bool)
```

Which fulfil the property that when the `z` in `term1` is equal to the
`y` in `term2`, the terms have the same result. As each term
introduces its own environment variable namespace, we cannot directly
check this.

What we want is a projection into a shared namespace, in which we can
reason about the equalities we desire. `mkobservation` and `refines`
can be parameterised with this projection, where `check` computes
them.