Add parameters for seed predicates.
Pleasingly, this has no effect on performance whatsoever, even though
n*num(properties) are now being considered.  Run-time is still
dominated by actually evaluating the terms, and the number of terms is
unchanged.

Because restriction to weakest preconditions is not yet implemented,
this finds a huge number of redundant properties.