Fix pure schema pruning / refactor generator
The problem was more subtle than I thought:

1. Expression type comparison was incorrect

(? :: Int) + (? :: Int)  `equivalent`  (? :: A)

returned true, for example, because the type check was not checking
the right thing, and then no assignments of type `A` were able to be
generated.  The conjunction of zero values is true, so they were
"equivalent"!

2. Result equivalence check was too restrictive AND too permissive

Too permissive in two ways: only the most general term of each schema
should be checked, and every assignment for a given renaming needs to
produce equal values.

Too restrictive in one way: only one renaming needs to be equivalent.

3. Pruning is really about terms, not schemas

The schema (? - ?) is made redundant in part by 0, but not in full.
Pruning is therefore really about terms, not schemas.  The generator
now still generates schemas, but also from each schema generates all
of its acceptable terms (these are in the annotation).

A term is kept if it is not: (a) pure and equivalent to another pure
term, or (b) has a subterm which has been forbidden in the past.

This adds 25 new properties to the semaphore output (and bumps the
execution time up to 17 minutes!)